+++ packages/viewer/web/src/RichContentViewer.tsx
@@ -69,7 +69,16 @@ export interface RichContentViewerProps {
 
 class RichContentViewer extends Component<
   RichContentViewerProps,
-  { raw?: RicosContent; error?: string }
+  {
+    raw?: RicosContent;
+    error?: string;
+    context: {
+      experiments?: AvailableExperiments;
+      isMobile: boolean;
+      t?: TranslationFunction;
+      containerWidth?: number;
+    };
+  }
 > {
   styles: Record<string, string>;
   typeMappers: PluginMapping;
@@ -92,7 +101,10 @@ class RichContentViewer extends Component<
     const styles = { ...viewerStyles, ...viewerAlignmentStyles, ...rtlStyle };
     this.styles = mergeStyles({ styles, theme: props.theme });
     this.typeMappers = combineMappers(props.typeMappers);
-    this.state = {};
+    const { experiments, isMobile = false, t, width } = props;
+    this.state = {
+      context: { experiments, isMobile, t, containerWidth: width },
+    };
   }
 
   static getInitialState = (props: RichContentViewerProps) => {
@@ -197,11 +209,7 @@ class RichContentViewer extends Component<
         inlineStyleMappers,
         locale,
         addAnchors,
-        isMobile = false,
-        t,
-        experiments,
         renderedInTable,
-        width,
       } = this.props;
       const wrapperClassName = classNames(styles.wrapper, {
         [styles.desktop]: !this.props.platform || this.props.platform === 'desktop',
@@ -240,7 +248,7 @@ class RichContentViewer extends Component<
       );
 
       return (
-        <GlobalContext.Provider value={{ experiments, isMobile, t, containerWidth: width }}>
+        <GlobalContext.Provider value={this.state.context}>
           <div className={wrapperClassName} dir={direction || getLangDir(locale)}>
             <div className={editorClassName}>{output}</div>
             <AccessibilityListener isMobile={this.props.isMobile} />