+++ CHANGELOG.md
@@ -16,12 +16,35 @@
     Click to see more.
   </summary>
 
+### :rocket: New Feature
+- `fullScreen`
+  - [#2206](https://github.com/wix-incubator/rich-content/pull/2206) fullscreen styles are bundled and so there's no need to import its `styles.min.css`
+
 ### :bug: Bug Fix
-- `plugin-link`
-  - [#2182](https://github.com/wix/ricos/pull/2182) open link panel from external toolbar in mobile
+- `common`
+  - [#2199](https://github.com/wix/ricos/pull/2199) bgColor fallback was undefined, influencing plugins
+### :house: Internal
+- `storybook`
+  - [#2191](https://github.com/wix/ricos/pull/2191) Theming - Custom Styles + refactor
 
 </details>
 <hr/>
+
+## 8.22.8 (Feb 23, 2021)
+### :bug: Bug Fix
+- `editor`
+  - [#2193](https://github.com/wix/ricos/pull/2193) fix content re-render issue
+- `viewer`
+  - [#2193](https://github.com/wix/ricos/pull/2193) fix content re-render issue
+
+
+## 8.22.7 (Feb 22, 2021)
+### :bug: Bug Fix
+- `plugin-link`
+  - [#2182](https://github.com/wix/ricos/pull/2182) open link panel from external toolbar in mobile
+### :house: Internal
+- `common`
+  - [#2189](https://github.com/wix/ricos/pull/2189) experiments: support non-boolean values
 
 ## 8.22.6 (Feb 22, 2021)
 ### :nail_care: Polish